shopper-concierge-demo/app/requirements.txt
gradio
pandas
google-generativeai
python-dotenv
google-cloud-aiplatform
